+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+
+## [0.0.4a] 03-04-2022
+
+### Bugfixes 
+- INTERACTION_ALREADY_REPLIED on /profile create exit
+It was happening because the reply on the start of the command works a return, and when the user finishes the creation of the profile a return have already happenend
+
+
 ## [0.0.4] 02-04-2022
 ### Added
 - /profile get ( fetch user info on CDD )